% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alencar2021.R
\name{alencar2021}
\alias{alencar2021}
\title{FD identificaiton method proposed by the Author (Pedro Alencar)}
\usage{
alencar2021(
  vtime,
  vprecipitation,
  vet0,
  crit = c(4, 2, 8, 2),
  months = c(3, 10)
)
}
\arguments{
\item{vtime}{data frame column or vector containing \code{date} data}

\item{vprecipitation}{data frame column or vector containing daily precipitation (mm.day-1)}

\item{vet0}{data frame column or vector containing daily potential evapotranspiration (Penman-Monteith - mm.day-1)}

\item{crit}{a vector with four elements indicating the thresholds (more in details).}

\item{months}{a vector with two elements, the first and last months in the analysis (definition of growing season)}
}
\value{
The function returns a list with two data frames. One with weekly and detailed values from the function and a second with a summary of all evetns identified.
}
\description{
FD identificaiton method proposed by the Author (Pedro Alencar)
}
\details{
The method tracks the variation (slope) of the precipitation and potential evapotranspiration.
It uses intra-year limits and the anomaly to the time series to identify exceptionally dry events.

The \code{crit} vector contains 4 elements:
    c1 = accumulation time for slope (in weeks) - Default = 4
    c2 = threshold for combined anomaly (sum) - Default = 2
    c3 = latency period of negative SPEI (in weeks) - Default = 8
    c4 = allows positive SPEI over latency period - Default = 2
}
\examples{
fd_Alencar <- alencar2021(vtime = df_d$time,
                          vprecipitation = df_d$precipitation,
                          vet0 = ET0$et0)


}
