% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Christian_clean_data_week.R
\name{Christian_clean_data_week}
\alias{Christian_clean_data_week}
\title{Function to clean the ESR data (necessary before running Christian et al. method)}
\usage{
Christian_clean_data_week(vtime, vET0, vETa, threshold = 1)
}
\arguments{
\item{vtime}{data frame column or vector containing \code{date} data}

\item{vET0}{data frame column or vector containing daily potential evapotranspiration (mm.day-1)}

\item{vETa}{data frame column or vector containing daily actual evapotranspiration (mm.day-1)}

\item{threshold}{a positive number, indicates the maximum ratio ETa/ET0 allowed.}
}
\value{
It returns a list with two elements. A data frame with time stamps and pentad ESR and a matrix with ESR data distributed in a 73 lines (pentads).
}
\description{
Function to clean the ESR data (necessary before running Christian et al. method)
}
\details{
Due to data and modelling quality, eventualy ETa >> ET0, causing errors, particularly in cold days when ET0 ≈ 0.
This function allows the reduction o distortions. Pentads discarded are interpolated.
}
\examples{
Christian2020_clean_data(vtime = df_d$time,
                         vET0 = ET0$et0, vETa = ETa$eta,
                         threshold = 1)

}
