% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_mat.R
\name{conf_matrix}
\alias{conf_matrix}
\title{Auxiliar (intern) function for calculating and exporting confusion matrix}
\usage{
conf_matrix(df, ref)
}
\arguments{
\item{df}{complete dataset with all events (by month)}

\item{ref}{the reference model - Mod. Ford and Labosier}
}
\value{
melted data frame to draw plot. The function return some confusion matrix common statistics:
  - Accuracy
  - Cohen's Kappa
  - Specificity
  - Sensitivity
  - Precision
  - F1
}
\description{
Auxiliar (intern) function for calculating and exporting confusion matrix
}
\details{
This function is supposed to be used only in the context of the
FD-Viz app
}
\examples{
all_fd <- process_all(de_tha_d,include_variables = T, data = 'station')$Series

confusion_matrix <- conf_matrix(all_fd, all_fd$`Osman et al.`)

}
