\name{de_tha_d}
\alias{de_tha_d}
\title{Daily data from FluxNET15 station Tharandt}

\description{
Dataframe from the original file. No alterations where made.

For more information about this dataset visit \url{http://sites.fluxdata.org/DE-Tha/}

\emph{Data under licence CC-BY-4.0}
}
\arguments{
\item{X}{a simple line counter}
\item{time}{time stamp)}
\item{precipitation}{in mm day-1}
\item{temperature}{in deg C}
\item{wind_speed}{in m s−1}
\item{pressure_atm}{in kPa}
\item{vapor_p_def}{in hPa}
\item{sensible_heat}{in W m-2}
\item{latent_heat}{in W m-2}
\item{soil_water}{soil water content in percentage}
}

\details{
For further information on the data also refer to Pastorello et al. (2020) - \url(https://doi.org/10.1038/s41597-020-0534-3)
}

