% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_tha_h.R
\docType{data}
\name{de_tha_h}
\alias{de_tha_h}
\title{Hourly data from FluxNET15 station Tharandt}
\format{
An object of class \code{data.frame} with 333120 rows and 9 columns.
}
\source{
\url{http://sites.fluxdata.org/DE-Tha/}
}
\usage{
de_tha_h
}
\description{

}
\keyword{datasets}
