% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanomaly.R
\name{f.anomaly}
\alias{f.anomaly}
\title{Function to calculate anomalies}
\usage{
f.anomaly(vector)
}
\arguments{
\item{vector}{A general vector or data frame column}
}
\value{
vector of anomalies.
}
\description{
Calculates the anomaly (number of standard deviations). This is an intern function.
}
