% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpercentile.R
\name{f.percentile}
\alias{f.percentile}
\title{Percentile function}
\usage{
f.percentile(vector)
}
\arguments{
\item{vector}{data frame column or vector containing the analysed data}
}
\value{
The function return a vector with percentiles
}
\description{
Calculates the percentile. This is an intern function.
}
\examples{

}
