% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fspei.R
\name{f.spei}
\alias{f.spei}
\title{Standard Precipitation Evaporation Index calculation}
\usage{
f.spei(vtime, vdeficit, n)
}
\arguments{
\item{vtime}{a data.frame column or vector with daily time stamps (Date type)}

\item{vdeficit}{a data.frame column or vector with daily hydrological deficit
obtained by the difference of precipitation and potential evapotranspitation (P - ET0)}

\item{n}{a natural number that indicates the accumulation time (pentad, week, month, etc)}
}
\value{
The function return a list with two elements. One data frame with time stamped pentad values and a matrix with years organized in columns.
}
\description{
Internal funciton to calculate the SPEI
}
