% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fyear.R
\name{f.year}
\alias{f.year}
\title{Separation by years - Internal function}
\usage{
f.year(i, day.var, year.var = NULL)
}
\arguments{
\item{i}{a numeric index}

\item{day.var}{data frame with time stamps (input)}

\item{year.var}{data frame with time stamps (output - default = \code{NULL})}
}
\value{
return a vector with all data pertaining to a year
}
\description{
Internal function to separate the data into years
}
