\name{f.percentile}
\alias{f.anomaly}
\title{Function obtain anomalies}
\usage{
f.anomaly(vector)
}
\description{
\emph{Intern function} to obtain anomalies from absolute data.
}

\arguments{
\item{vector}{A vector or dataframe's column or row with numbers to obtain anomalies (\emph{z-score}) distribution from.}
}

\examples{
f.anomaly()
}
