\name{fd_fl2017}
\alias{fd_fl2017}
\title{Function - Flash Drought classification via Foster and Labosier 2017}
\usage{
fd_fl2017(data)
}
\description{
Function of classify as FD based on Foster and Labosier2017.
}

\arguments{
\item{data}{a data.frame with two columns, time and soil water content (SWC) respectivelly}

}
\examples{
fd_fl2017(df)
}
