\name{flx_de_tha_hour}
\alias{flx_de_tha_hour}
\title{Hourly data from FluxNET15 station Tharandt}

\description{
  Dataframe of the original file. No alterations where made.

  For more information about this dataset visit \url{http://sites.fluxdata.org/DE-Tha/}

  \emph{Data under licence CC-BY-4.0}
}

