% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FordLabosier_gs.R
\name{FordLabosier_gs}
\alias{FordLabosier_gs}
\title{Modified method of Ford and Labosier (2017)}
\usage{
FordLabosier_gs(vtime, vswc, crit = c(40, 20, 30))
}
\arguments{
\item{vtime}{data frame column or vector containing \code{date} data}

\item{vswc}{data frame column or vector containing soil water content values}

\item{crit}{a vector of three value (default \code{c(20,40,30)}) indicating the model thresholds for lower, upper and persistance limits for SWC percentiles.}
}
\value{
A list with two data frames, one a time series with all data for FD identification, and the second with a summary of FD events.
}
\description{
Besides the additional duration criteria (at least 3 pentads under 30th p.)
we also added two new criteria:
   1. Events with onset out of the growing season (MAMJJASO) are not considered
      as flash droughts
   2. Evetns that are 3 pentads or less apart are considers the same event.
}
\examples{
FD_events <- FordLabosier_modif(de_tha_d$time, de_tha_d$soil_water, crit = c(40,20,30))


}
