% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df_era5.R
\name{get_df_era5}
\alias{get_df_era5}
\title{get_df_era5 - Generate DF from ERA5 raw (netcedf) data}
\usage{
get_df_era5(list_files, lat, lon, soil_layers = c(1))
}
\arguments{
\item{list_files}{vector containing the name of file(s)}

\item{lat}{latitude of study area (decimal degrees)}

\item{lon}{longitude of study area (decimal degrees)}

\item{soil_layers}{vector with on or more values between 1 and 4. This indicates which are the soil layers desired in the analysis. More information in the function description.}
}
\value{
The function returns a data frame with all variables of interest:
        Precipitation (tp)
        Temperature (t2m)
        Potential evapotranspiration (pev)
        actual evapotranspiration (e)
        volumetric soil water content (vswl1, vswl2, vswl3, vswl4)
}
\description{
The ERA5 provides 4 soil water content layers:
        Layer 1: 0 - 7cm,
        Layer 2: 7 - 28cm,
        Layer 3: 28 - 100cm,
        Layer 4: 100 - 289cm.
}
