% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_fluxnet.R
\name{get_df_fluxnet}
\alias{get_df_fluxnet}
\title{get_df_fluxnet - Generate DF from FLUXNET raw data}
\usage{
get_df_fluxnet(filename, timestep, soil_level = 1)
}
\arguments{
\item{filename}{name of fluxnet file}

\item{timestep}{what is the analysed timestep, if daily or hourly}

\item{soil_level}{Which is the sensor index (depends on the station, please pay attention to the .csv file)}
}
\value{

}
\description{
Builds data frame to be used as input in all functions from the raw FLUXNET data (.csv)
}
\examples{

}
