% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicriteria_fd.R
\name{multicriteria_fd}
\alias{multicriteria_fd}
\title{Multiple criteria method of FD classification}
\usage{
multicriteria_fd(
  vtime,
  vtemp,
  vprec,
  vet0,
  veta,
  score = 0.6,
  d_score = 0.2,
  thresholds = c(1, 0, 0, 0, -2, 50, 10, 30)
)
}
\arguments{
\item{vtime}{data frame column or vector containing \code{date} data}

\item{vtemp}{data frame column or vector containing daily temperature}

\item{vprec}{data frame column or vector containing daily precipitation}

\item{vet0}{data frame column or vector containing daily ET0 (potential
evapotranspiration). It can be obtained with the function \code(penman_day)}

\item{veta}{data frame column or vector containing daily ETa (actual
evapotranspiration). It can be obtained with the function \code(actual_evap_day)}

\item{score}{a number ranging from 0 to 100. The percentile above which
events will be considered flash droughts.}

\item{thresholds}{a vector with multiple threshold values. View details for
more information. We advise not to change from the default values.}
}
\value{
The function returns a list with two data frames. One with weekly and detailed values from the function and a second with a summary of all events identified.
}
\description{
Multiple criteria method of FD classification
}
\details{
The \code{thresholds} vector contains 8 elements respectively:
\itemize{
\item 
\enumerate{
\item The threshold used for cleaning data in Christian et al. method
(\code{Christian_clean_data_week}). It indicates the maximum value accepted
for the ration ETa/ET0. The default is 1.1, allowing ETa marginally higher than
ET0. This reduces significatly the number o missing data.
}
\item 
\enumerate{
\item General threshold for anomalies. Indicates the number of standard deviations
to the average for the anomalies. The default value is 1.
}
\item 
\enumerate{
\item SPEI threshold. The value to be considered under drought conditions. The
default value is the same froim Noguera et al. (2020), i.e., -1.28
}
\item 
\enumerate{
\item General threshold for accumulated anomalies. Indicated the sum of weekly
anomalies accumulated over a period of 4 weeks. The default value is 3.
}
\item 
\enumerate{
\item SESR and SPEI intensifications. The accumulated difference over 4 weeks.
The default value is -2.
}
\item 6 and 7) EDDI intensification and recuperation. The values are the same from
\code{Pendergrass2020}. Default: 50 and 10 respectively
\item 
\enumerate{
\item Delta_SESR percentile. Indicates the percentile of the difference over 4
weeks. We used the same threshold from Christian et al. 2020 for default (30th percentile.)
}
}
}
\examples{
ET0 <- penman_day(vtime = de_tha_d$time, vwind = de_tha_d$wind_speed, vtemp = de_tha_d$temperature,
                  vvpd = de_tha_d$vapor_p_def, vheatflux = (de_tha_d$sensible_heat + de_tha_d$latent_heat))[,2]

ETa <- actual_evap_day(vtime = de_tha_d$time,vlatent_heat =de_tha_d$latent_heat, vtemperature = de_tha_d$temperature)[,2]

multi_criteria_fd <- multicriteria_fd(vtime = de_tha_d$time, vtemp = de_tha_d$temperature,
                                      vprec = de_tha_d$precipitation, vet0 = ET0, veta = ETa)


}
