\name{precipiration_day}
\alias{precipiration_day}
\title{function to extract data from a pixel in a NC file for an expecific variable}
\usage{
precipiration_day(data.prec)
}

\description{
*Auxiliar function - get subdaily precipitation data into days.
}

\arguments{
\item{data.prec}{dataframe with two columns, timestamp and precipitation (in milimiters)}
}

\examples{
precipiration_day()
}
