% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparenc.R
\name{prepare.nc}
\alias{prepare.nc}
\title{Title}
\usage{
prepare.nc(data, period = "day", f = function(x) x)
}
\arguments{
\item{data}{data frame with time stamps}

\item{period}{string with interval of accumulation (default = 'day')}

\item{f}{\code{R function} to be applied (mean, max, sum, etc)}
}
\value{

}
\description{
Title
}
\examples{

}
