% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actual_evap_day.R
\name{actual_evap_day}
\alias{actual_evap_day}
\title{Calculate daily actual evaporation}
\usage{
actual_evap_day(vtime, vlatent_heat, vtemperature = 20)
}
\arguments{
\item{vtime}{data frame column or vector containing \code{date} data}

\item{vlatent_heat}{data frame column or vector containing daily latent heat flux (W.m-2.day-1)}

\item{vtemperature}{data frame column or vector containing daily average tempeature (Celcius)}
}
\value{
Data frame containing dates (daily) and evaporation (mm/day)
}
\description{
Calculate daily actual evaporation
}
\details{
The calculation is based on the description in "A short course in Cloud Physics" (Rogers and Yau, 1989)
}
\examples{
ETa <- actual_evap_day(vtime = de_tha_d$time,
                       vlatent_heat =de_tha_d$latent_heat,
                       vtemperature = de_tha_d$temperature)



}
