% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Christian2020.R
\name{Christian2020}
\alias{Christian2020}
\title{FD identification method from Christian et al. (2020)}
\usage{
Christian2020(esr_list)
}
\arguments{
\item{esr_list}{The list output from the function \code{Christian2020_clean_data}}
}
\value{
The function returns a list with two data frames. One with weekly and detailed values from the function and a second with a summary of all evetns identified.
}
\description{
FD identification method from Christian et al. (2020)
}
\examples{
#' fd_Christian <- Christian2020_clean_data(vtime = df_d$time,
vET0 = ET0$et0, vETa = ETa$eta,
threshold = 2) \%>\% Christian2020()

}
