% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_tha_d.R
\docType{data}
\name{de_tha_d}
\alias{de_tha_d}
\title{Daily data from FluxNET15 station Tharandt}
\format{
An object of class \code{data.frame} with 6940 rows and 9 columns.
}
\source{
\url{http://sites.fluxdata.org/DE-Tha/}
}
\usage{
de_tha_d
}
\description{

}
\keyword{datasets}
