% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddi_anomaly.R
\name{eddi}
\alias{eddi}
\title{Calculate EDDI}
\usage{
eddi(vet0)
}
\arguments{
\item{vet0}{data frame column or vector containing daily ET0 (potential evapotranspiration). It can be obtained with the function \code(penman_day)}
}
\value{
vector with EDDI values
}
\description{
function to calculate EDDI (Evaporative Demand Drought Index) following Hobbins et al., 2016
}
\examples{
eddi_values <- penman_day(vtime = de_tha_d$time, vwind = de_tha_d$wind_speed,
                          vvpd = de_tha_d$vapor_p_def, vtemp = de_tha_d$temperature,
                          vheatflux = (de_tha_d$sensible_heat + de_tha_d$latent_heat))[,2] \%>\%
                          eddi()

}
