% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddi_percentile.R
\name{eddi_percentile}
\alias{eddi_percentile}
\title{Function to calculate EDDI in percentiles on a week accumulation time}
\usage{
eddi_percentile(vtime, vet0, dist = "tukey")
}
\arguments{
\item{vtime}{a data.frame column or vector with daily time stamps (Date type)}

\item{vet0}{a data.frame column or vector with ordered daily ET0 values. They
can be obtained directly from reanalysis/models or from the functions
\code{penman_day} or \code{hargreaves_day}}

\item{dist}{string containing the distribution used to calibrate EDDI. It can
either be 'tukey' following the original method from Hobbings et al (2016)
or log-logist' according to Noguera et al. (2021)}
}
\value{

}
\description{
Function to calculate EDDI in percentiles on a week accumulation time
}
\examples{
ET0 <- penman_day(vtime = df_d$time, vwind = df_d$wind_speed,
                  vvpd = df_d$vapor_p_def, vtemp = df_d$temperature,
                  vheatflux = (df_d$sensible_heat + df_d$latent_heat))

percentiles_eddi <- eddi_percentile(vtime = de_tha_d$time, ET0$et0, dist = 'tukey')

}
