% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hargreaves_day.R
\name{hargreaves_day}
\alias{hargreaves_day}
\title{Hargreaves-Samany daily-ET0 function}
\usage{
hargreaves_day(vtime, vtemperature, my_lat)
}
\arguments{
\item{vtime}{data frame column or vector containing \code{date} data}

\item{vtemperature}{data frame column or vector containing temperature}

\item{my_lat}{Latitude of study area}
}
\value{
The function return a data frame with two columns, time stamps (daily) and ET0
}
\description{
Hargreaves-Samany daily-ET0 function
}
\examples{

}
