% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mo2016.R
\name{Mo2016}
\alias{Mo2016}
\title{FD identification based on Mo & Lettenmeier (2015, 2016)}
\usage{
Mo2016(
  vtime,
  vprecipitation,
  vtemperature,
  vsoil_water,
  vlatent_heat = NULL,
  vevap = NULL,
  flux_data = T
)
}
\arguments{
\item{vtime}{data frame column or vector containing \code{date} data}

\item{vprecipitation}{data frame column or vector containing daily precipitation}

\item{vtemperature}{data frame column or vector containing daily temperature}

\item{vsoil_water}{data frame column or vector containing soil water content (soil moisture)}

\item{vlatent_heat}{data frame column or vector containing daily total latent heat flux (to calculate actual evapotranspiration)}

\item{vevap}{data frame column or vector containing daily actual evapontranspiration (if already available)}

\item{flux_data}{boolean, indicates if the data is from eddy covariance stations.}
}
\value{
The function returns a list with two data frames. One with pentad and detailed values from the function and a second with a summary of all events identified.
}
\description{
FD identification based on Mo & Lettenmeier (2015, 2016)
}
\examples{
fd_Mo <- Mo2016(vtime = df_d$time, vprecipitation = df_d$precipitation,
                vtemperature = df_d$temperature, vsoil_water = df_d$soil_water,
                vlatent_heat = df_d$latent_heat, flux_data = T)

}
