% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Noguera2020.R
\name{Noguera2020}
\alias{Noguera2020}
\title{FD identification based on Noguera et al. (2020)}
\usage{
Noguera2020(vtime, vprecipitation, vet0, threshold = NA)
}
\arguments{
\item{vtime}{data frame column or vector containing \code{date} data}

\item{vprecipitation}{data frame column or vector containing daily precipitation}

\item{vet0}{data frame column or vector containing daily ET0 (potential evapotranspiration). It can be obtained with the function \code(penman_day)}

\item{threshold}{numeric, indicate the lower limit to identify a FD (see description)}
}
\value{
The function returns a list with two data frames. One with weekly and detailed values from the function and a second with a summary of all events identified.
}
\description{
The function uses the SPEI to identify a FD. The intentisication period is defined by a 4 week (one month) period with a SPEI reduction of 2. The SPEI value must go equal or below the threshold.
If \code{threshold} is not provided -1.28 (10 % quantile in a normal distribution) is assigned.
}
