% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Osman2021.R
\name{Osman2021}
\alias{Osman2021}
\title{FD identification based on Osman et al. (2020)}
\usage{
Osman2021(vtime, vswc, threshold = 20)
}
\arguments{
\item{vtime}{data frame column or vector containing \code{date} data}

\item{vswc}{data frame column or vector containing soil moisture data}

\item{threshold}{a numeric value (default = 20) indicating the lower limit of SM percentile for FD identification}
}
\value{
The function returns a list with two data frames. One with pentad and detailed values from the function and a second with a summary of all events identified.
}
\description{
FD identification based on Osman et al. (2020)
}
\examples{
fd_Osman  <- Osman2021(vtime = df_d$time,
                       vswc = df_d$soil_water,
                       threshold = 20)

}
