% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pendergrass2020.R
\name{Pendergrass2020}
\alias{Pendergrass2020}
\title{FD identification by Pendergrass et al. (2021)}
\usage{
Pendergrass2020(vtime, vet0, limit.down = 10)
}
\arguments{
\item{vtime}{a data.frame column or vector with daily time stamps (Date type)}

\item{vet0}{a data.frame column or vector with ordered daily ET0 values. They
can be obtained directly from reanalysis/models or from the functions
\code{penman_day} or \code{hargreaves_day}}

\item{limit.down}{a numeric value, indicating the limit of recuperation after
onset. This criterion is additional, to flexibilise the original method.
We set it's value to 10 as default. TO run the original method (more restrict)
define it to 0.}
}
\value{
Function \code{Pendergrass2020} retuns a list with two data frames.
1) a complete time stamped series containing relevant variables and FD events;
2) a summary of each event with its duration and interval of occurance.
}
\description{
Identifies Flash Drougt evetns using EDDI variations, as described in
Pendergrass et al. "Flash droughts present a new challenge for
subseasonal-to-seasonal prediction", 2021.
}
\details{
This function is based on EDDI calculated according to Hobbings et al. (2016)
DOI: 10.1175/jhm-d-15-0121.1
It uses a simple empirical Tukey plotting position to assess EDDI percentiles.
}
\examples{
fd_Pendergrass <- Pendergrass2020(vtime = de_tha_d$time,
                                  vet0 = ET0$et0, limit.down = 10)

}
