% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penman_day.R
\name{penman_day}
\alias{penman_day}
\title{penman_day - Daily ET0 using Penman-Monteith}
\usage{
penman_day(vtime, vwind, vtemp, vvpd, vheatflux, altitude = 0)
}
\arguments{
\item{vtime}{data frame column or vector containing \code{date} data}

\item{vwind}{data frame column or vector containing daily wind velocity}

\item{vtemp}{data frame column or vector containing daily temperature}

\item{vvpd}{data frame column or vector containing daily vapor pressure}

\item{vheatflux}{data frame column or vector containing daily total heat flux (latent and sensible heat flux)}

\item{altitude}{study area altitude in meters above sea level}
}
\value{
The function returns a data frame with time stamped daily potential evapotranspiration
}
\description{
penman_day - Daily ET0 using Penman-Monteith
}
\examples{
 ET0 <- penman_day(vtime = de_tha_d$time, vwind = de_tha_d$wind_speed,
                   vvpd = de_tha_d$vapor_p_def, vtemp = de_tha_d$temperature,
                   vheatflux = (de_tha_d$sensible_heat + de_tha_d$latent_heat))

}
