% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_all.R
\name{process_all}
\alias{process_all}
\title{Process all methods with default values}
\usage{
process_all(df_d, include_variables = T, data = "station")
}
\arguments{
\item{df_d}{a data frame with all relevant variables. It can be obtained using
the functions \code{get_df_era5} or \code{get_df_fluxnet}}

\item{include_variables}{Boolean,informs with all variables should be included
in the output}

\item{data}{string value, either 'station' or 'reanalysis'.}
}
\value{
The function returns a list with two data frames. One with daily and detailed values for all methods and a list with one data frame for each method, each containing a summary of all identified events.
}
\description{
Process all methods with default values
}
\details{
The function run all methods with their default thresholds
}
\examples{
  all_methods <- process_all(de_tha_d,include_variables = T, data = 'station')
}
